EbayAutolister/requirements.txt
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
cryptography>=3.4.0
click>=8.0.0
pyyaml>=6.0
tqdm>=4.64.0
openai>=1.51.0
openai-agents>=0.1.0
